# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SKN API
  version: 1.0.0-hackathon
  description: |
    React 모바일 앱과 Spring Boot API가 실제로 사용하는 계약이다.
    인증은 HttpOnly 세션 쿠키를 사용하고 OpenAI API 키는 서버에서만 읽는다.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /me/routines/current:
    get:
      tags: [Routines]
      summary: 현재 실제 사용 루틴을 조회한다
      operationId: getCurrentRoutine
      responses:
        '200': { description: 현재 루틴, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [Routines]
      summary: 제품과 순서를 저장하고 새 사용 경험을 시작한다
      operationId: replaceCurrentRoutine
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRoutineRequest' }
      responses:
        '200': { description: 새 루틴 경험, content: { application/json: { schema: { $ref: '#/components/schemas/Experience' } } } }
        '409': { $ref: '#/components/responses/Problem' }
  /me/routines/{routineId}:
    get:
      tags: [Routines]
      summary: 저장된 루틴 버전 하나의 구성을 조회한다
      operationId: getRoutine
      parameters: [ { $ref: '#/components/parameters/RoutineId' } ]
      responses:
        '200': { description: 루틴 구성, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /me/routines/{routineId}/name-suggestion:
    post:
      tags: [Routines]
      summary: 루틴 이름과 개인 데이터 기반 AI 문장을 준비한다
      description: 제품 종류·순서·시간·빈도로 이름을 제안하고, 관련 개인 기록과 자기보고 선호가 있으면 이 루틴이 사용 경험에 어떤 도움을 줄지 짧은 한 문장으로 생성해 저장된 루틴 버전에 연결한다. AI 실패 시에도 서버가 구성 기반 기본 이름을 반환하며 루틴 저장이나 이름 적용을 막지 않는다.
      operationId: suggestRoutineName
      parameters: [ { $ref: '#/components/parameters/RoutineId' } ]
      responses:
        '200': { description: 이름 제안, content: { application/json: { schema: { $ref: '#/components/schemas/RoutineNameSuggestion' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /me/routines/{routineId}/name:
    put:
      tags: [Routines]
      summary: 사용자가 확인한 이름을 현재 루틴과 진행 중인 경험에 적용한다
      operationId: renameCurrentRoutine
      parameters: [ { $ref: '#/components/parameters/RoutineId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RenameRoutineRequest' }
      responses:
        '200': { description: 이름이 반영된 루틴, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /me/routines/{routineId}/insight:
    post:
      tags: [Routines]
      summary: 저장된 루틴의 개인 데이터 기반 AI 도움 문장을 준비한다
      description: 이미 생성된 문장이 있으면 그대로 반환하고, 없으면 관련 개인 기록과 자기보고 선호에서 생성해 해당 루틴 버전에 저장한다. 루틴의 제품·순서·이름은 바꾸지 않는다.
      operationId: generateRoutineInsight
      parameters: [ { $ref: '#/components/parameters/RoutineId' } ]
      responses:
        '200': { description: 읽기 전용 루틴 AI 문장, content: { application/json: { schema: { $ref: '#/components/schemas/RoutineInsight' } } } }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /me/routines/baseline:
    get:
      tags: [Routines]
      summary: Rescue 비교에 쓰는 마지막 문제없던 루틴을 조회한다
      operationId: getBaselineRoutine
      responses:
        '200': { description: 비교 기준 루틴, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
...
    RoutineId: { name: routineId, in: path, required: true, schema: { type: integer, format: int64 } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1 } }
  responses:
    Problem:
      description: RFC 9457 문제 응답
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Product:
      type: object
      required: [id, brand, name, category, verified, guide, facts, personalRecordCount, owned]
      properties:
        id: { type: integer, format: int64 }
        brand: { type: string }
        brandLogoUrl: { type: [string, 'null'], description: 저장소에 등록된 브랜드 로고 애셋 경로. 실제 제조 법인 정보가 아님 }
        name: { type: string }
        category: { type: string }
        volume: { type: [string, 'null'] }
        versionLabel: { type: [string, 'null'] }
        verified: { type: boolean, description: legacy 확인 flag와 source-backed fact가 모두 있을 때만 true }
        guide: { $ref: '#/components/schemas/ProductGuide' }
        facts: { type: array, items: { $ref: '#/components/schemas/ProductFact' } }
        personalRecordCount: { type: integer, minimum: 0 }
        owned: { type: boolean }
        imageUrl: { type: [string, 'null'], format: uri }
    ProductGuide:
      type: object
      required: [summary, routineStep, usageType, usageTiming, usageInstructions, highlights, origin, generatedAt]
      properties:
        summary: { type: string, minLength: 1, maxLength: 500, description: 제품명과 등록 제형을 포함해 제품 종류를 설명하는 짧은 안내 }
        routineStep: { type: string, minLength: 1, maxLength: 100, description: category를 바탕으로 한 일반적인 루틴 위치 }
        usageType: { type: string, minLength: 1, maxLength: 100, description: 씻어내기·붙이기·남겨두기 등 일반 사용 형태 }
        usageTiming:
          type: array
          maxItems: 4
          items: { type: string, minLength: 1, maxLength: 100 }
        usageInstructions:
          type: array
          description: 제품 표시를 우선하도록 안내하는 일반 사용법. 출처 확인 제품 사실이 아님
          maxItems: 4
          items: { type: string, minLength: 1, maxLength: 300 }
        highlights:
          type: array
          description: 등록 제형을 첫 항목으로 포함해 제품 종류와 사용 형태를 이해하기 위한 특징 설명. 관찰·기록 과제가 아님
          maxItems: 4
          items: { $ref: '#/components/schemas/ProductHighlight' }
        origin: { type: string, enum: [AI_GENERATED, EDITORIAL] }
        generatedAt: { type: string, format: date-time }
    ProductHighlight:
      type: object
      required: [title, detail]
      properties:
        title: { type: string, minLength: 1, maxLength: 80 }
        detail: { type: string, minLength: 1, maxLength: 300 }
    ProductFact:
      type: object
      required: [type, text, sourceLabel, sourceUrl, checkedAt]
      properties:
        type:
          type: string
          enum: [DIRECTIONS, TEXTURE, LABEL_CLAIM, CAUTION, SUN_PROTECTION, INGREDIENT_LABEL, CERTIFICATION]
        text: { type: string, minLength: 1, maxLength: 500 }
        sourceLabel: { type: string, minLength: 1, maxLength: 150 }
        sourceUrl: { type: string, format: uri }
        checkedAt: { type: string, format: date-time }
...
    UserProduct:
      type: object
      required: [id, addedAt, personalRecordCount, inCurrentRoutine]
      properties:
        id: { type: integer, format: int64 }
        product: { anyOf: [ { $ref: '#/components/schemas/Product' }, { type: 'null' } ] }
        customBrand: { type: [string, 'null'], maxLength: 120 }
        brandLogoUrl: { type: [string, 'null'], description: 카탈로그 또는 정확히 일치하는 직접 입력 브랜드의 로고 애셋 경로 }
        customName: { type: [string, 'null'], maxLength: 160 }
        customCategory: { type: [string, 'null'], maxLength: 80 }
        memo: { type: [string, 'null'], maxLength: 1000 }
        addedAt: { type: string }
        personalRecordCount: { type: integer, minimum: 0 }
        inCurrentRoutine: { type: boolean }
...
    RoutineItem:
      type: object
      required: [userProductId, productName, brand, category, timeSlot, position, frequency]
      properties:
        userProductId: { type: integer, format: int64 }
        productName: { type: string }
        brand: { type: string }
        brandLogoUrl: { type: [string, 'null'] }
        category: { type: string }
        timeSlot: { type: string, enum: [MORNING, EVENING, BOTH] }
        position: { type: integer, minimum: 0 }
        frequency: { type: string }
    RoutineItemInput:
      type: object
      required: [userProductId, timeSlot, frequency]
      properties:
        userProductId: { type: integer, format: int64 }
        timeSlot: { type: string, enum: [MORNING, EVENING, BOTH] }
        frequency: { type: string, maxLength: 20 }
    Routine:
      type: object
      required: [id, name, dayPart, status, startedAt, items]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        dayPart: { type: string, enum: [MORNING, EVENING, ANYTIME] }
        status: { type: string, enum: [CURRENT, BASELINE, PAST] }
        startedAt: { type: string }
        items: { type: array, items: { $ref: '#/components/schemas/RoutineItem' } }
        insight: { $ref: '#/components/schemas/RoutineInsight' }
    RoutineInsight:
      type: object
      required: [text, keywords, generatedAt]
      properties:
        text: { type: string, minLength: 1, maxLength: 70, description: 이 루틴이 사용 경험에 어떤 도움을 줄지 개인 기록과 자기보고 선호에서 AI가 만든 읽기 전용 한 문장 }
        keywords: { type: array, minItems: 0, maxItems: 3, items: { type: string, minLength: 1, maxLength: 30 }, description: 같은 개인 맥락에서 정리한 짧은 표현 2~3개. 기존 생성분은 보강 전까지 빈 배열일 수 있음 }
        generatedAt: { type: string }
    UpdateRoutineRequest:
      type: object
      required: [name, items]
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        items: { type: array, minItems: 1, maxItems: 12, items: { $ref: '#/components/schemas/RoutineItemInput' } }
    RoutineNameSuggestion:
      type: object
      required: [name, aiGenerated]
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        aiGenerated: { type: boolean }
        insight: { $ref: '#/components/schemas/RoutineInsight' }
    RenameRoutineRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
...
    Experience:
      type: object
      required: [id, subjectType, title, subtitle, status, startedAt, reviewDueAt, day, daysUntilReview, reviewDue, recordSummary]
      properties:
        id: { type: integer, format: int64 }
        subjectType: { type: string, enum: [PRODUCT, ROUTINE] }
        routineId: { type: [integer, 'null'], format: int64 }
        userProductId: { type: [integer, 'null'], format: int64 }
        title: { type: string }
        subtitle: { type: string }
        status: { type: string, enum: [ACTIVE, COMPLETED, CANCELLED] }
        startedAt: { type: string }
        reviewDueAt: { type: string }
        day: { type: integer, minimum: 1 }
        daysUntilReview: { type: integer, minimum: 0 }
        reviewDue: { type: boolean }
        routine: { anyOf: [ { $ref: '#/components/schemas/Routine' }, { type: 'null' } ] }
        product: { anyOf: [ { $ref: '#/components/schemas/UserProduct' }, { type: 'null' } ] }
        latestRecord: { anyOf: [ { $ref: '#/components/schemas/ExperienceRecord' }, { type: 'null' } ] }
        recordSummary: { $ref: '#/components/schemas/ExperienceRecordSummary' }
    ExperienceRecordSummary:
      type: object
      description: 같은 사용 경험 기간에 누적된 개별 기록의 표시용 집계. 원본 기록을 대체하지 않는다.
      required: [totalCount, likedCount, disappointedCount, unsureCount, discomfortCount]
      properties:
        totalCount: { type: integer, minimum: 0 }
        likedCount: { type: integer, minimum: 0 }
        disappointedCount: { type: integer, minimum: 0 }
        unsureCount: { type: integer, minimum: 0 }
        discomfortCount: { type: integer, minimum: 0 }
...
    ExperienceRecord:
      type: object
      required: [id, productName, sentiment, note, discomfort, adherence, tags, createdAt]
      properties:
        id: { type: integer, format: int64 }
        sessionId: { type: [integer, 'null'], format: int64 }
        userProductId: { type: [integer, 'null'], format: int64 }
        productName: { type: string }
        sentiment: { type: string, enum: [LIKED, DISAPPOINTED, UNSURE] }
        note: { type: string }
        discomfort: { type: string, enum: [NOT_REPORTED, REPORTED, UNKNOWN] }
        adherence: { type: string, enum: [MATCHED, PARTIAL, DIFFERENT, UNKNOWN] }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string }
...
    Problem:
      type: object
      required: [type, title, status, detail, code, retryable]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        retryable: { type: boolean }